+++ benchmarks/__init__.py
@@ -1,1 +1,8 @@
-#  noqa
+"""ASV benchmark suite for MAICoS."""
+
+import sys
+from pathlib import Path
+
+_REPO_ROOT = Path(__file__).resolve().parents[1]
+if str(_REPO_ROOT) not in sys.path:
+    sys.path.insert(0, str(_REPO_ROOT))
