+++ benchmarks/core/__init__.py
@@ -1,0 +1,1 @@
+"""Benchmarks for :mod:`maicos.core`."""
